Third and Fourth Features added

+++ app/api/listing_routes.py
@@ -1,12 +1,12 @@
 from flask import Blueprint, jsonify, request, redirect
-from app.models import db, Listing, User, Review
+from app.models import db, Listing, User, Review, MenuItem
 from ..forms.listing_form import ListingForm
-
+from ..forms.menu_item_form import MenuItemForm
 from ..forms.review_form import ReviewForm
 from datetime import date
 
 from flask_login import current_user, login_required
-
+from .aws_helpers import get_unique_filename, upload_file_to_s3, remove_file_from_s3
 
 listing_routes = Blueprint('listing', __name__)
 
@@ -96,6 +96,7 @@ def get_owned_listings():
 
     return { "listings": owned_listings }
 
+#! Create Listing ///////////////////////////////////////////////////////////////////////////
 
 @listing_routes.route('/', methods=["POST"])
 @login_required
@@ -110,6 +111,17 @@ def create_listing():
 
     if form.validate_on_submit():
 
+        image_url = form.data["image_url"]
+        image_url.filename = get_unique_filename(image_url.filename)
+        upload = upload_file_to_s3(image_url)
+
+        print(upload)
+
+        print ("url")
+
+        if "url" not in upload:
+            return { "errors": "Error uploading image to S3" }, 400 #! check if this is correct?
+
         new_listing = Listing(
             owner_id=current_user.id, #! this is where the owner id is coming from not from the front end
             address=form.data["address"],
@@ -121,7 +133,10 @@ def create_listing():
             price=form.data["price"],
             open_hours=form.data["open_hours"],
             close_hours=form.data["close_hours"],
-            image_url=form.data["image_url"],
+
+            # image_url=form.data["image_url"],
+            image_url=upload["url"],
+
             created_at = date.today(),
             updated_at = date.today()
         )
@@ -132,9 +147,9 @@ def create_listing():
     else:
         print(form.errors)
         return { "errors": form.errors }, 400
+#! //////////////////////////////////////////////////////////////////////////////////////////////////
 
-
-@listing_routes.route("/<int:listingId>", methods=["PUT"])
+@listing_routes.route("/<int:listingId>", methods=["PUT"]) #! I need to find how to update the AWS image here / I did update it with AWS need to check if it works
 @login_required
 def update_listing(listingId):
     """
@@ -145,10 +160,29 @@ def update_listing(listingId):
     form["csrf_token"].data = request.cookies["csrf_token"]
 
     listing_to_update = Listing.query.get(listingId)
-    print( "listing_to_update =======>>" , listing_to_update)
-    print( "current_user.id +++++++++>>" , current_user.id)
+    # print( "listing_to_update =======>>" , listing_to_update)
+    # print( "current_user.id +++++++++>>" , current_user.id)
+
     if listing_to_update.owner_id == current_user.id:
+        # Delete associated S3 files
+        remove_file_from_s3(listing_to_update.image_url)
+
         if form.validate_on_submit():
+            image = form.data["image_url"]
+            image.filename = get_unique_filename(image.filename)
+
+            # Upload the image to S3
+            upload = upload_file_to_s3(image)
+            print(upload)
+
+            if 'url' not in upload:
+                return { "errors": "Error uploading image to S3" }, 400
+
+            # Use the S3 URL
+            image_url = upload['url']
+
+
+
             listing_to_update.address = form.data["address"]
             listing_to_update.city = form.data["city"]
             listing_to_update.state = form.data["state"]
@@ -158,11 +192,13 @@ def update_listing(listingId):
             listing_to_update.price = form.data["price"]
             listing_to_update.open_hours = form.data["open_hours"]
             listing_to_update.close_hours = form.data["close_hours"]
-            listing_to_update.image_url = form.data["image_url"]
+            listing_to_update.image_url = image_url
+
             db.session.commit()
             return listing_to_update.to_dict()
 
         else:
+            print(form.errors)
             return { "errors": form.errors }, 400
 
     else:
@@ -173,15 +209,22 @@ def update_listing(listingId):
 @login_required
 def delete(listingId):
     """
-    Delete a listing
+    Delete a listing and associated S3 files
     """
     listing_to_delete = Listing.query.get(listingId)
 
     if listing_to_delete:
         if listing_to_delete.owner_id == current_user.id:
-            db.session.delete(listing_to_delete)
-            db.session.commit()
-            return { "message": "Delete successful!" }
+
+            file_delete =remove_file_from_s3(listing_to_delete.image_url)
+            print("file delete", file_delete)
+
+            if file_delete is True:
+                db.session.delete(listing_to_delete)
+                db.session.commit()
+                return { "message": "Delete successful!" }
+            else:
+                return {"error": "failed to delete the file from S3"}, 500
         else:
             return { "message": "FORBIDDEN" }, 403
     else:
@@ -227,6 +270,65 @@ def create_review(listingId):
         db.session.commit()
         return new_review.to_dict(), 201
 
+    else:
+        print(form.errors)
+        return { "errors": form.errors }, 400
+
+#! MENU ITEM ////////////////////////////////////////////////
+
+@listing_routes.route('/<int:listingId>/menuitems')
+#/api/listings/:listingId/menuitems
+def get_listing_menu_items(listingId):
+    """
+    Query for all menu items for a specific listing
+    """
+
+    all_menu_items = MenuItem.query.all()
+
+    listing_menu_items = [menu_item.to_dict() for menu_item in all_menu_items if menu_item.listingId == listingId]
+
+    return listing_menu_items
+
+@listing_routes.route('/<int:listingId>/createmenuitem', methods=["POST"])
+#/api/listings/:listingId/createmenuitem
+@login_required
+def create_menu_item(listingId):
+    """
+    Route to post a new menu item
+    """
+    form = MenuItemForm()
+    form["csrf_token"].data = request.cookies["csrf_token"]
+
+    if form.validate_on_submit():
+        image = form.data["imageUrl"]
+        image.filename = get_unique_filename(image.filename)
+
+        # Upload the image to S3
+        upload = upload_file_to_s3(image)
+        print(upload)
+
+        if 'url' not in upload:
+            return { "errors": "Error uploading image to S3" }, 400
+
+        # Use the S3 URL
+        image_url = upload['url']
+
+        new_menu_item = MenuItem(
+            listingId=listingId,
+            name=form.data["name"],
+            size=form.data["size"],
+            calories=form.data["calories"],
+            price=form.data["price"],
+            description=form.data["description"],
+            # Use the S3 URL
+            imageUrl=image_url,
+            created_at = date.today(),
+            updated_at = date.today()
+        )
+
+        db.session.add(new_menu_item)
+        db.session.commit()
+        return new_menu_item.to_dict(), 201
     else:
         print(form.errors)
         return { "errors": form.errors }, 400
